Calculating heating demand in the highest possible DG
This script is calculating the heating demand in the highest possible DG for each object. This is done by joining standard_aufbereitet with waermebedarfswerte_bis_dg3 on the combination of building typology and baualtersklasse. Then the related wärmebedarf is multiplied with the gebäudenutzfläche of the object. The result is the heating demand